+++ src/Resources/assets/js/pwa_public_build.js
@@ -1,0 +1,3 @@
+import "core-js/stable";
+import "regenerator-runtime/runtime";
+import './contao-pwa-bundle';