+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_content'];
+
+$lang['pwaSubscribeButtonTemplate'] = ["Button-Template", 'Template fpr den "Benachrichtigungen abonnieren"-Button auswählen.'];